+++ mne/realtime/fieldtrip_client.py
@@ -333,10 +333,9 @@ class FieldTripClient(object):
         Parameters
         ----------
         stop_measurement : bool
-            Also stop the measurement.
+            unused, for compatibility.
         """
-        if self._recv_thread is not None:
-            self._recv_thread = None
+        self._recv_thread = None
 
     def iter_raw_buffers(self):
         """Return an iterator over raw buffers.